fix cumprod and cumsum frontend for JAX.

+++ ivy/functional/frontends/jax/lax/operators.py
@@ -152,8 +152,8 @@ def convert_element_type(operand, new_dtype):
 
 def cumprod(operand, axis=0, reverse=False):
     if reverse:
-        return ivy.flip(ivy.cumprod(ivy.flip(operand), axis))
-    return ivy.cumprod(operand, axis)
+        return ivy.flip(ivy.cumprod(ivy.flip(operand), axis, dtype=operand.dtype))
+    return ivy.cumprod(operand, axis, dtype=operand.dtype)
 
 
 cumprod.unsupported_dtypes = {"torch": ("float16",)}
@@ -161,8 +161,8 @@ cumprod.unsupported_dtypes = {"torch": ("float16",)}
 
 def cumsum(operand, axis=0, reverse=False):
     if reverse:
-        return ivy.flip(ivy.cumsum(ivy.flip(operand), axis))
-    return ivy.cumsum(operand, axis)
+        return ivy.flip(ivy.cumsum(ivy.flip(operand), axis, dtype=operand.dtype))
+    return ivy.cumsum(operand, axis, dtype=operand.dtype)
 
 
 cumsum.unsupported_dtypes = {"torch": ("float16",)}
